Lesson 7
Add text with v-if directive that shows when no items are present. Comment out items. Wrap header in a div. Add two buttons to header div. Add a state property to data, set to default. Add v-if directive to hide form div if state  is default. Add v-if and v-else directives to new buttons to show add item button when state is default and cancel button when state is edit. Add change state method. Call change state method on new button click to change state.