Lesson 9
Refactor items to be an array of objects to allow for tracking purchased and priority. Update li variable, and saveItem method for new structure. Add method to toggle purchased state. Add v-on to li to call method for togglePurchased. Add v-bind to add strikeout class if item purchased is true. Add style for strikeout and priority classes to css.